Force reflectance values to be >= 0 in NDVI calculation
The built in .normalizedDifference() function returns no data if either of the inputs are negative, but many of the water bodies have negative reflectance values in collection 2 and get masked out.

The NDVI and other indices may need to be adjusted or overwritten for the water masking to work correctly.