Moved water_mask calculation inside landsat module
The WRI, NDWI, and MNDWI inputs to the Image class can all be removed since they are only used to generate the water mask.

Changed calculation to use either the QA_PIXEL water mask or the built in water_mask (with and added max albedo limit to prevent saturated pixels getting flagged as water).